backend/api_foodgram/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-filter==22.1
djangorestframework==3.13.1
flake8==5.0.4
mccabe==0.7.0
Pillow==9.2.0
pkg_resources==0.0.0
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.2.1
sqlparse==0.4.2
